requirements.txt
pydantic>=2.6.1
python-dotenv>=1.0.0
openai>=1.12.0
pandas>=2.2.0
llama-parse>=0.0.40
python-multipart>=0.0.9
tqdm>=4.66.1
pathlib>=1.0.1